simple phone number searching
If you enter a numeric search query it is treated as part of a phone number.
Because ConTagged has to deal with any possible formatting in the phone number
entries and the limits of LDAP search filters it will match any number where
the given digits occur in the same order (but not nesserally besides each others)

Eg. a search for 3940 will match the number +49 30 22 49 46 30

darcs-hash:20070627120322-6e07b-6c934b3f883a8cb9f84f171e2f7670727c51bdf0.gz