embed 'clippy' to copy address to clipboard
It is often convinient to copy a complete address to the system
clipboard; to achieve this, this patch includes the flash widget
"clippy" to the contact view. Clicking on one of the widgets copies
either the office or home address to the clipboard, making it simple to
include it in other applications.

Clippy itself is hosted here: https://github.com/mojombo/clippy